Two references broke the rules that were written to stop this, in opposite directions
The goal names two ways a reference can be unfair: `Iterator<T>` or `Stream`
where the subject boxes nothing, and a field narrower than the f64 a TypeScript
`number` is. `user-iterable` broke both.

`bytes/op` says this lane allocates **zero** per step, so the `Iterator<Integer>`
was measuring `Integer` against our codegen. And `Steps` had `int at` and
`int seed` where field specialization does not narrow, so this lane carries
doubles and calls `NtsRuntime.toInt` twice a step for the two `| 0`s -- the
reference was getting a proof this compiler has not made.

The corrections pull opposite ways: dropping the box makes the reference faster,
widening the fields makes it slower. Both are required, which is the argument
for the rules being written down rather than decided per row.

Widening the fields immediately found a bug in my own comment. It asserted that
`(int)` and `| 0` agree here "because `seed` is masked to 255" -- reading the
mask on the *return* as though it were on the field. `seed` is a full int32,
`seed * 31` reaches 2^36, and a cast saturates where `ToInt32` wraps. The
checksum said so at once. With `int` fields the wrap is the hardware's and the
mistake is unavailable, so the reference could not have shown that this compiler
is calling a runtime helper twice a step to get it.

Then I audited the rest against `javap` rather than against memory:

    in-narrowing   emits  5 int      ref int      fair
    symbol-keys    emits  2 int 2 Z  ref matching fair
    upcast         emits  3 int      ref int      fair
    instanceof     emits  3 double   ref int      VIOLATION

`instanceof` never masks, so specialization has nothing to prove an i32 from and
the fields stay `D`. Its reference used `int` -- a narrowing this compiler did
not make. Now `double`, checked against `javap`.

That one is worth naming because of its direction: it made **us** look worse,
not better. Nobody auditing for flattery would have found it. The rule is not
"always f64", it is "whatever the subject actually carries", and the only way to
know which is to look at the class file.

+++ benches/cases/instanceof/ref.java
@@ -13,37 +13,51 @@
 // not a default. That matters for the test order: `Circle` and `Square` both
 // extend `Shape`, so asking about `Shape` first would swallow all three.
 //
-// The fields are `int` because the source stays inside int32, and each subclass
-// adds its own on top of the base's, which is what `Layout.base` records and
-// what makes `super_class` real here rather than a prefix coincidence.
+// **The fields are `double`, and that is checked rather than assumed.** `javap`
+// on what this backend emits says `nts.gen.Shape.size`, `Circle.radius` and
+// `Square.side` are all `D`: the source never masks, so specialization has
+// nothing to prove an i32 from and the fields stay f64. An `int` here would
+// hand the reference a narrowing this compiler did not make, which is the
+// "no field narrower than the f64 a TypeScript `number` is" rule -- and it
+// would do it in the direction that makes *us* look worse, so it was not going
+// to be caught by anyone checking for flattery.
+//
+// The sibling case `upcast` masks with `| 0` and `& 0xffff`, specialization
+// takes it, and `javap` there says `I`. Its reference uses `int` for the same
+// reason this one does not. The rule is not "always f64"; it is "whatever the
+// subject actually carries", and the only way to know is to look.
+//
+// Each subclass adds its own field on top of the base's, which is what
+// `Layout.base` records and what makes `super_class` real here rather than a
+// prefix coincidence.
 final class Ref {
     static class Shape {
-        int size;
+        double size;
 
-        Shape(int size) {
+        Shape(double size) {
             this.size = size;
         }
     }
 
     static final class Circle extends Shape {
-        int radius;
+        double radius;
 
-        Circle(int size) {
+        Circle(double size) {
             super(size);
             this.radius = size + 1;
         }
     }
 
     static final class Square extends Shape {
-        int side;
+        double side;
 
-        Square(int size) {
+        Square(double size) {
             super(size);
             this.side = size + 2;
         }
     }
 
-    private static Shape shape(int i) {
+    private static Shape shape(double i) {
         if (i % 3 == 0) {
             return new Circle(i);
         }
